Create README.md
These two functions allow you to define a matrix, x, and calculate its inverse.  If the inverse already exists, the inverse is called cache. 
The first function, makeCacheMatrrix, allows you to set the values of a matrix, get a previously defined matrix and get and set an inverse of this matrix, if it can be successfully inverted.